doc tests

+++ src/types/enums.rs
@@ -1,34 +1,34 @@
 
 /// For quickly defining single-character-alpha enum variants:
-/// '''
-/// nsdq-util::define_enum!{
+/// ```
+/// use nsdq_util::define_enum;
+///
+/// define_enum!{
 ///     MyEnum:
 ///         "This is my strong type for a character-based protocol enum.";
-///
+/// 
 ///     ['A'] VariantA
 ///         "This is the first variant for MyEnum.",
 ///     ['B'] VariantB
 ///         "This is the other variant for MyEnum."
 /// }
 ///
-/// fn test() {
+/// let bytes = b"AB";
 ///
-///     let bytes = b"AB";
+/// let (bytes, parsed_1) = MyEnum::parse(bytes).unwrap();
+/// let (bytes, parsed_2) = MyEnum::parse(bytes).unwrap();
 ///
-///     let (bytes, parsed_1) = MyEnum::parse(bytes).unwrap();
-///     let (bytes, parsed_2) = MyEnum::parse(bytes).unwrap();
+/// assert_eq!(parsed_1, MyEnum::VariantA);
+/// assert_eq!(parsed_2, MyEnum::VariantB);
+/// assert!(bytes.is_empty());
 ///
-///     assert_eq!(parsed_1, MyEnum::VariantA);
-///     assert_eq!(parsed_2, MyEnum::VariantB);
-///     assert!(bytes.is_empty());
+/// let mut bytes = vec![];
+/// bytes.extend(MyEnum::VariantA.encode());
+/// bytes.extend(MyEnum::VariantB.encode());
 ///
-///     let mut bytes = vec![];
-///     bytes.extend(MyEnum::VariantA.encode());
-///     bytes.extend(MyEnum::VariantB.encode());
+/// assert_eq!(&bytes, b"AB");
 ///
-///     assert_eq!(&bytes, b"AB");
-/// }
-/// '''
+/// ```
 #[macro_export] macro_rules! define_enum {
     ($name:ident: $edoc:expr; $([$tag:expr] $kind:ident $kdoc:expr),*$(,)?) => {
 
@@ -58,7 +58,7 @@
 
             pub fn encode(&self) -> [u8; 1] {
                 match self {$(
-                    $name::$kind => [$tag],
+                    $name::$kind => [$tag as u8],
                 )*}
             }
         }
